refactor: export gql fucntions to separate file
